improved roles, teacher can see all theses etc

+++ app/Http/Controllers/InternshipController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+#use Helper;
 use App\Internship;
 use Auth;
 
@@ -24,9 +25,28 @@ class InternshipController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        //
+        $user_id = Auth::user()->id;
+        $user = User::find($user_id);
+        if ($user->hasRole('Õpetaja'))
+        {
+            if($request->input('company_name'))
+            {
+                $internships_all = Internship::where('company_name', 'like', '%'.$request->input('company_name').'%')->get();
+            }
+            else
+            {
+                $internships_all = Internship::all();
+            }
+            
+            return view('internships')
+                ->with('internships_all', $internships_all);
+        }
+        else 
+        {
+            return redirect('/home');
+        }
     }
 
     /**
@@ -36,7 +56,17 @@ class InternshipController extends Controller
      */
     public function create()
     {
-        return view('user.create_internship');
+        $user_id = Auth::user()->id;
+        $user = User::find($user_id);
+        //check if current user has student role
+        if($user->hasRole('Õpilane'))
+        {
+            return view('user.create_internship');
+        }
+        else 
+        {
+            return redirect('/home');
+        }
     }
 
     /**
@@ -47,7 +77,13 @@ class InternshipController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
+        $user_id = Auth::user()->id;
+        $user = User::find($user_id);
+        //check if current user has student role
+        
+        if ($user->hasRole('Õpilane'))
+        {
+            $this->validate($request, [
             'company_name' => 'required|string|max:50',
             'start_date' => 'required|date',
             'start_date' => 'required|date',
@@ -55,7 +91,6 @@ class InternshipController extends Controller
             ]);
             
              //Create new internship for current user
-             
              $internship = new Internship;
              $internship->user_id = Auth::user()->id;
              $internship->company_name = $request->input('company_name');
@@ -65,6 +100,12 @@ class InternshipController extends Controller
              $internship->save();
              
              return redirect('/home')->with('success', 'Lõputöö lisatud!');
+        }
+        else 
+        {
+            return redirect('/home');
+        }
+        
     }
 
     /**
@@ -87,12 +128,21 @@ class InternshipController extends Controller
     public function edit($id)
     {
         $internship = Internship::find($id);
-        
-        if(Auth::user()->id !== $internship->user_id){
-            return redirect('/home')->with('error', 'Teil puudub ligipääs!');
+        $user_id = Auth::user()->id;
+        $user = User::find($user_id);
+        if ($user->hasRole('Õpilane'))
+        {
+            if(Auth::user()->id !== $internship->user_id)
+            {
+                return redirect('/home')->with('error', 'Teil puudub ligipääs!');
+            }
+            
+            return view ('user.edit_internship')->with('internship', $internship);
         }
-        
-        return view ('user.edit_internship')->with('internship', $internship);
+        else 
+        {
+            return redirect('/home');
+        }
     }
 
     /**
@@ -104,7 +154,11 @@ class InternshipController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
+        $user_id = Auth::user()->id;
+        $user = User::find($user_id);
+        if ($user->hasRole('Õpilane'))
+        {
+            $this->validate($request, [
             'company_name' => 'required|string|max:50',
             'start_date' => 'required|date',
             'start_date' => 'required|date',
@@ -112,7 +166,6 @@ class InternshipController extends Controller
             ]);
             
              //Update internship for current user
-             
              $internship = Internship::find($id);
              $internship->company_name = $request->input('company_name');
              $internship->start_date = $request->input('start_date');
@@ -121,6 +174,11 @@ class InternshipController extends Controller
              $internship->save();
              
              return redirect('/home')->with('success', 'Praktika andmed muudetud!');
+        }
+        else 
+        {
+            return redirect('/home');
+        }
     }
 
     /**
@@ -132,12 +190,21 @@ class InternshipController extends Controller
     public function destroy($id)
     {
         $internship = Internship::find($id);
+        $user_id = Auth::user()->id;
+        $user = User::find($user_id);
         
-        if(Auth::user()->id !== $internship->user_id){
-            return redirect('/home')->with('error', 'Teil puudub ligipääs!');
+        if ($user->hasRole('Õpilane'))
+        {
+            if(Auth::user()->id !== $internship->user_id){
+                return redirect('/home')->with('error', 'Teil puudub ligipääs!');
+            }
+            
+            $internship->delete();
+            return redirect('/home');
         }
-        
-        $internship->delete();
-        return redirect('/home');
+        else 
+        {
+            return redirect('/home');
+        }
     }
 }
